+++ macfiles-native/src/main/kotlin/garden/ephemeral/macfiles/native/NativeCommon.kt
@@ -8,7 +8,7 @@ import java.text.Normalizer
 internal fun getVolumePath(file: File): File {
     // Find the filesystem
     val st = SystemB.Statfs()
-    SystemB.INSTANCE.statfs(file.absolutePath.toByteArray(), st)
+    SystemB.INSTANCE.statfs64(file.absolutePath.toByteArray(), st)
     // File and folder names in HFS+ are normalized to a form similar to NFD.
     // Must be normalized (NFD->NFC) before use to avoid unicode string comparison issues.
     return File(Normalizer.normalize(st.f_mntonname.decodeToString().trim('\u0000'), Normalizer.Form.NFC))